Use AsyncStream<Never> for waitUnlessCancelled
withUnsafeContinuation needs a withTaskCancellationHandler to support cooperative cancellation. A simpler approach is to use the continuation by an AsyncStream. AsyncStream has a built in cancellation support. An element type of Never guarantees that there will be no emits. So the stream can only be cancelled or finished.

See https://www.pointfree.co/episodes/ep210-clocks-controlling-time